move context serialization and type it better
