+++ src/serialize-spec.ts
@@ -12,11 +12,9 @@ import {
 import { assert, mapObject } from './util.ts';
 import { Credential } from './credential-index.ts';
 import type { InputJSON, NodeJSON, SpecJSON } from './validation.ts';
-import type { HttpsInputContext, ZkAppInputContext } from './context.ts';
 
 export {
   type SerializedValue,
-  type SerializedContext,
   serializeNode,
   deserializeNode,
   serializeInput,
@@ -24,8 +22,6 @@ export {
   serializeSpec,
   deserializeSpec,
   validateSpecHash,
-  serializeInputContext,
-  deserializeInputContext,
 };
 
 function serializeSpec(spec: Spec): SpecJSON {
@@ -261,73 +257,6 @@ function deserializeNode(root: any, node: NodeJSON): Node {
       node satisfies never;
       throw Error(`Invalid node type: ${type}`);
   }
-}
-
-type SerializedContext =
-  | { type: 'https'; action: string; serverNonce: SerializedValue }
-  | {
-      type: 'zk-app';
-      action: string;
-      serverNonce: SerializedValue;
-      verifierIdentity: {
-        publicKey: SerializedValue;
-        tokenId: SerializedValue;
-        network: 'mainnet' | 'devnet' | { custom: string };
-      };
-    }
-  | null;
-
-function serializeInputContext(
-  context: undefined | ZkAppInputContext | HttpsInputContext
-): SerializedContext {
-  if (context === undefined) return null;
-  let serverNonce = serializeProvable(context.serverNonce);
-  if (context.type === 'https') {
-    return { type: context.type, serverNonce, action: context.action };
-  }
-  if (context.type === 'zk-app') {
-    let { publicKey, tokenId, network } = context.verifierIdentity;
-    return {
-      type: context.type,
-      serverNonce,
-      action: context.action,
-      verifierIdentity: {
-        publicKey: serializeProvable(publicKey),
-        tokenId: serializeProvable(tokenId),
-        network,
-      },
-    };
-  }
-  throw Error(
-    `Unsupported context type: ${(context satisfies never as any).type}`
-  );
-}
-function deserializeInputContext(
-  context: SerializedContext
-): undefined | ZkAppInputContext | HttpsInputContext {
-  if (context === null) return undefined;
-  if (context.type === 'https') {
-    return {
-      type: context.type,
-      action: context.action,
-      serverNonce: deserializeProvable(context.serverNonce),
-    };
-  }
-  if (context.type === 'zk-app') {
-    return {
-      type: context.type,
-      action: context.action,
-      serverNonce: deserializeProvable(context.serverNonce),
-      verifierIdentity: {
-        publicKey: deserializeProvable(context.verifierIdentity.publicKey),
-        tokenId: deserializeProvable(context.verifierIdentity.tokenId),
-        network: context.verifierIdentity.network,
-      },
-    };
-  }
-  throw Error(
-    `Unsupported context type: ${(context satisfies never as any).type}`
-  );
 }
 
 async function hashSpec(serializedSpec: string): Promise<string> {
